+++ pkg/output/options.go
@@ -18,7 +18,6 @@ type Options struct {
 
 	WebhookOptions
 	GCPPubsubOptions
-	HTTPServerOptions
 }
 
 type WebhookOptions struct {
@@ -33,12 +32,4 @@ type GCPPubsubOptions struct {
 	Topic        string // Topic name. Will create it if not exists.
 	Subscription string // Subscription name. Will create it if not exists.
 	Clear        bool   // Clear will clear all topics and subscriptions before running.
-}
-
-type HTTPServerOptions struct {
-	TLSCertificate  string        // TLS certificate file path.
-	TLSKey          string        // TLS key file path.
-	ResponseHeaders []string      // KV list of response headers.
-	ReadTimeout     time.Duration // HTTP Server read timeout.
-	WriteTimeout    time.Duration // HTTP Server write timeout.
 }